pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "GSAutoML"
version = "0.0.1"
authors = [
  { name="Data_Team", email="youssef.mahmoud@gizasystems.com" }
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
        "psycopg2==2.9.6",
        "python-decouple==3.6",
        "numpy==1.24.2",
        "pandas==1.3.5",
        "statsmodels==0.14.0",
        "scikit-learn==1.2.2",
        "prophet==1.1.4",
        "xgboost==1.7.6",
        "lightgbm==4.0.0",
        "tpot==0.12.1",
        "smac>=0.13.0"
]